requirements.txt
fastapi[all]==0.115.0
pydantic==2.9.2
pydantic[email]
uvicorn==0.31.0
jinja2==3.1.4
pydantic_settings==2.5.2
aiosqlite==0.20.0
alembic==1.13.3
SQLAlchemy==2.0.35
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
loguru==0.7.2
markdown2==2.5.1